src/oracle/game/requirements.txt
flask
pycryptodome
uwsgi
